PHP 8.1 | Tests: temporarily ignore select test failures [2]
The test failures being handled by setting an expectation for a PHP deprecation notice are on the list to be fixed via the "Filter extension"-CS action list.

Once the particular issue affecting these tests has been fixed, these expectations can be removed (and this is safeguarded by the expectation as the test will start to fail because the deprecation will no longer occur).